requirements/conda.txt
# conda installables
fusepy
gevent
paramiko
setuptools
